fix: server-render blog listing page for SEO and crawler visibility
Blog listing page was fully client-rendered ("use client"), making all
blog posts invisible to search engines, social link previews, and
crawlers like Google. Static fetchers see "Loading blog..." instead of
actual content.

Split into server component + client component:
- Server component fetches featured posts, categories, and initial posts
  via ConvexHttpClient at request time
- Client component receives server data as props, hydrates with real-time
  Convex subscriptions for live updates
- Graceful fallback if NEXT_PUBLIC_CONVEX_URL is unavailable (client-only)
- Uses existing blog-transforms utilities for type mapping
- Preserves all interactive features (search, filtering, pagination)
